He estado realizando las operaciones de CRUD y alguna prueba unitaria que aun no comprendo del todo

+++ SpringBoot/Futbol_v1/src/main/java/com/futbol/model/Equipo.java
@@ -22,10 +22,9 @@ public class Equipo {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "idequipo")
-	// Lo usare para que aparezca en la documentacion
 	@Schema(name = "id", description = "Identificador único para el equipo", example = "1", required = true)
-	private Long idequipo;
-	
+	private Integer idequipo;
+
 	@NotEmpty(message = "El nombre del equipo es obligatorio")
 	@Column(name = "nombre")
 	private String nombre;
